move file(s) to new directory
